Fix effect authority CI fallout
